Struct mls_crypto_provider::MlsCryptoProvider

pub struct MlsCryptoProvider { /* private fields */ }

Implementations

impl MlsCryptoProvider

pub async fn try_new_with_configuration( config: MlsCryptoProviderConfiguration<'_>, ) -> MlsProviderResult<Self>

Initialize a CryptoProvider with a backend following the provided config (see: MlsCryptoProviderConfiguration)

pub async fn try_new( db_path: impl AsRef<str>, identity_key: impl AsRef<str>, ) -> MlsProviderResult<Self>

pub async fn try_new_in_memory( identity_key: impl AsRef<str>, ) -> MlsProviderResult<Self>

pub fn new_with_store( key_store: CryptoKeystore, entropy_seed: Option<EntropySeed>, ) -> Self

Initialize a CryptoProvided with an already-configured backing store

pub async fn update_pki_env( &self, pki_env: PkiEnvironment, ) -> MlsProviderResult<()>

Replaces the PKI env currently in place

pub async fn is_pki_env_setup(&self) -> bool

Returns whether we have a PKI env setup

pub fn reseed(&mut self, entropy_seed: Option<EntropySeed>)

Reseeds the internal CSPRNG entropy pool with a brand new one.

If None is provided, the new entropy will be pulled through the current OS target’s capabilities

pub async fn close(self) -> MlsProviderResult<()>

Closes this provider, which in turns tears down the backing store

Note: This does not destroy the data on-disk in case of persistent backing store

pub async fn destroy_and_reset(self) -> MlsProviderResult<()>

Tears down this provider and obliterates every single piece of data stored on disk.

you have been warned

pub fn borrow_keystore(&self) -> &CryptoKeystore

Borrow keystore

pub fn borrow_keystore_mut(&mut self) -> &mut CryptoKeystore

pub fn unwrap_keystore(self) -> CryptoKeystore

Allows to retrieve the underlying key store directly